project-root/requirements.txt
langchain
streamlit
python-dotenv
google-generativeai
PyMuPDF
